backend/app/services/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-text-splitters
langchainhub==0.1.21
langsmith==0.3.8
macholib==1.16.3
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.7
openai==1.63.0
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
playwright==1.50.0
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.1.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tk==0.1.0
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0 